# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend modular API for GNyx. (Powered by Kubex Ecosystem)",
        "title": "GNyx API",
        "termsOfService": "https://gnyx.kubex.world/terms",
        "contact": {
            "name": "GNyx API Support",
            "url": "https://gnyx.kubex.world",
            "email": "contact@kubex.world"
        },
        "version": "1.0.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/v1/callback": {
            "get": {
                "description": "Exchanges the Google authorization code, issues local tokens, and redirects back to the frontend.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Handle Google OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Frontend redirect path",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect back to frontend",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.authErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.authErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}